+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 🐞 Bug fixes
 
 - Properly check ImageBitmap ([#2942](https://github.com/maplibre/maplibre-gl-js/pull/2942), [#2940](https://github.com/maplibre/maplibre-gl-js/issues/2940))
+- VectorTileWorkerSource: fix reload for original's load parse would not pass the rawTileData and meta. ([#2941](https://github.com/maplibre/maplibre-gl-js/pull/2941))
 - _...Add new stuff here..._
 
 ## 3.2.1